Day 18A Implementation as BigInteger

+++ src/main/java/org/jgoeres/adventofcode2020/Day18/RunDay18.java
@@ -1,23 +1,26 @@
 package org.jgoeres.adventofcode2020.Day18;
+
+import java.math.BigInteger;
+
 public abstract class RunDay18 {
     static String pathToInputs = "data/day18/input.txt";
 
     static Day18Service day18Service = new Day18Service();
 
-    public static long problem18A() {
+    public static BigInteger problem18A() {
         System.out.println("=== DAY 18A ===");
 
-        long result = day18Service.doPartA();
+        BigInteger result = day18Service.doPartA();
 
         System.out.println("Day 18A: Answer = " + result);
 
         return result;
     }
 
-    public static long problem18B() {
+    public static BigInteger problem18B() {
         System.out.println("=== DAY 18B ===");
 
-        long result = day18Service.doPartB();
+        BigInteger result = day18Service.doPartB();
 
         System.out.println("Day 18B: Answer = " + result);
